pyproject.toml
[project]
name = "orak-2025-starter-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "ahk>=1.8.4",
    "ahk-binary>=2025.1.0",
    "aiohttp>=3.13.2",
    "anthropic>=0.72.0",
    "backoff>=2.2.1",
    "chardet>=5.2.0",
    "colorama>=0.4.6",
    "dacite>=1.9.2",
    "dataclass-wizard>=0.35.1",
    "dill>=0.4.0",
    "dxcam>=0.0.5",
    "easyocr>=1.7.2",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.49.0",
    "gym==0.26.2",
    "gym-super-mario-bros>=7.4.0",
    "gymnasium==1.1.1",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "hydra-core>=1.3.2",
    "imageio[ffmpeg]>=2.37.2",
    "langchain-chroma>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-google-vertexai>=3.2.0",
    "loguru>=0.7.0",
    "matplotlib>=3.10.7",
    "mss>=10.1.0",
    "multi-template-matching==1.6.6",
    "nes-py==8.2.1",
    "nest-asyncio>=1.6.0",
    "numpy==1.26.4",
    "omegaconf>=2.3.0",
    "openai>=2.7.1",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "pillow>=10.3.0",
    "pyautogui>=0.9.54",
    "pyboy==2.5.2",
    "pydantic>=2.12.3",
    "pydirectinput>=1.0.4 ; sys_platform == 'win32'",
    "pygame>=2.6.1",
    "pyobjc-framework-cocoa>=12.0 ; sys_platform == 'darwin'",
    "pyobjc-framework-quartz>=12.0 ; sys_platform == 'darwin'",
    "python-dotenv>=1.2.1",
    "pywin32>=311 ; sys_platform == 'win32'",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-image>=0.25.2",
    "segment-anything",
    "spacy>=3.8.7",
    "supervision>=0.26.1",
    "tenacity>=9.1.2",
    "termcolor>=3.2.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "typer>=0.20.0",
    "ultralytics>=8.3.225",
    "burnysc2==7.1.1",
    "screeninfo>=0.8.1",
    "wandb>=0.23.1",
    "weave>=0.52.22",
]

[tool.uv]
# burnysc2==7.1.1 declares `numpy>=2.0.0`, but this project intentionally pins numpy==1.26.4.
# `override-dependencies` tells uv to prefer this version even if it would otherwise be considered
# an invalid resolution, matching pip's "install anyway" behavior.
override-dependencies = [
    "numpy==1.26.4",
    # burnysc2==7.1.1 declares `protobuf<4.0.0` but other dependencies (e.g. grpcio-tools) require
    # newer protobuf releases. We intentionally prefer a modern protobuf here.
    "protobuf==6.33.2",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
